refactor: comment 기능 패키지 분리
