+++ android/app/src/main/java/com/digginroom/digginroom/data/di/ViewModelFactory.kt
@@ -4,9 +4,9 @@ import android.content.Context
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewmodel.initializer
 import androidx.lifecycle.viewmodel.viewModelFactory
+import com.digginroom.digginroom.feature.room.customview.roominfoview.comment.CommentViewModel
 import com.digginroom.digginroom.feature.join.JoinViewModel
 import com.digginroom.digginroom.feature.login.LoginViewModel
-import com.digginroom.digginroom.feature.room.CommentViewModel
 import com.digginroom.digginroom.feature.room.RoomViewModel
 import com.digginroom.digginroom.feature.scrap.ScrapRoomViewModel
 import com.digginroom.digginroom.feature.scrap.ScrapViewModel
