+++ android/app/src/main/java/com/digginroom/digginroom/feature/room/customview/roominfoview/RoomInfoView.kt
@@ -4,6 +4,7 @@ import android.content.Context
 import android.view.LayoutInflater
 import androidx.constraintlayout.widget.ConstraintLayout
 import com.digginroom.digginroom.databinding.RoomInfoBinding
+import com.digginroom.digginroom.feature.room.customview.roominfoview.comment.listener.ShowCommentsListener
 import com.digginroom.digginroom.feature.room.RoomEventListener
 import com.digginroom.digginroom.feature.room.customview.roominfoview.navigator.DefaultRoomNavigator
 import com.digginroom.digginroom.model.CommentModel
