--- android/app/src/main/java/com/digginroom/digginroom/feature/room/customview/roominfoview/ShowCommentsListener.kt
@@ -1,6 +1,0 @@
-package com.digginroom.digginroom.feature.room.customview.roominfoview
-
-interface ShowCommentsListener {
-
-    fun show(roomId: Long)
-}
