--- android/app/src/main/java/com/digginroom/digginroom/feature/room/customview/roominfoview/ShowRoomCommentListener.kt
@@ -1,8 +1,0 @@
-package com.digginroom.digginroom.feature.room.customview.roominfoview
-
-import com.digginroom.digginroom.model.CommentModel
-
-interface ShowRoomCommentListener {
-
-    fun show(trackModel: List<CommentModel>)
-}
