+++ android/app/src/main/java/com/digginroom/digginroom/feature/room/customview/roominfoview/comment/adapter/CommentAdapter.kt
@@ -1,8 +1,8 @@
-package com.digginroom.digginroom.feature.room.customview.roominfoview.adapter
+package com.digginroom.digginroom.feature.room.customview.roominfoview.comment.adapter
 
 import android.view.ViewGroup
 import androidx.recyclerview.widget.ListAdapter
-import com.digginroom.digginroom.feature.room.customview.roominfoview.ShowCommentMenuListener
+import com.digginroom.digginroom.feature.room.customview.roominfoview.comment.listener.ShowCommentMenuListener
 import com.digginroom.digginroom.model.CommentModel
 
 class CommentAdapter : ListAdapter<CommentModel, CommentViewHolder>(CommentDiffUtilCallback()) {
