+++ android/app/src/main/java/com/digginroom/digginroom/feature/room/customview/roominfoview/comment/adapter/CommentDiffUtilCallback.kt
@@ -1,4 +1,4 @@
-package com.digginroom.digginroom.feature.room.customview.roominfoview.adapter
+package com.digginroom.digginroom.feature.room.customview.roominfoview.comment.adapter
 
 import androidx.recyclerview.widget.DiffUtil
 import com.digginroom.digginroom.model.CommentModel
