+++ android/app/src/main/java/com/digginroom/digginroom/feature/room/customview/roominfoview/comment/listener/CommentEventListener.kt
@@ -1,4 +1,4 @@
-package com.digginroom.digginroom.feature.room.customview.roominfoview
+package com.digginroom.digginroom.feature.room.customview.roominfoview.comment.listener
 
 interface CommentEventListener {
     fun findComments()
