+++ android/app/src/main/java/com/digginroom/digginroom/feature/room/customview/roominfoview/comment/listener/CommentMenuEventListener.kt
@@ -1,0 +1,6 @@
+package com.digginroom.digginroom.feature.room.customview.roominfoview.comment.listener
+
+interface CommentMenuEventListener {
+    fun update()
+    fun delete()
+}
