+++ android/app/src/main/java/com/digginroom/digginroom/feature/room/customview/roominfoview/comment/listener/ShowCommentsListener.kt
@@ -1,0 +1,6 @@
+package com.digginroom.digginroom.feature.room.customview.roominfoview.comment.listener
+
+interface ShowCommentsListener {
+
+    fun show(roomId: Long)
+}
